evans: next 14 + migrate static mdx files to rsc
