+++ apps/admin/package.json
@@ -27,8 +27,8 @@
     "lodash.isequal": "^4.5.0",
     "lodash.omit": "^4.5.0",
     "lodash.set": "^4.3.2",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-router-dom": "^6.3.0",
     "uuid": "^8.3.2"
   },
@@ -41,8 +41,8 @@
     "@types/lodash.isequal": "^4.5.5",
     "@types/lodash.omit": "^4.5.6",
     "@types/lodash.set": "^4.3.6",
-    "@types/react": "^18.0.22",
-    "@types/react-dom": "^18.0.7",
+    "@types/react": "18.2.67",
+    "@types/react-dom": "18.2.22",
     "@types/react-router-dom": "^5.3.3",
     "@types/uuid": "^8.3.1",
     "@vitejs/plugin-react": "3.1.0",
