+++ apps/cli/package.json
@@ -69,6 +69,7 @@
     "@types/lodash.isequal": "^4.5.5",
     "@types/node-fetch": "^2.5.7",
     "@types/puppeteer-core": "^2.1.0",
+    "@types/react": "18.2.67",
     "@types/sharp": "^0.28.2",
     "@types/uuid": "^3.3.0",
     "@types/yargs": "^15.0.0"
