+++ apps/cover-web-app/package.json
@@ -28,17 +28,17 @@
     "classnames": "^2.2.6",
     "glob": "^7.1.4",
     "lodash": "^4.17.11",
-    "react": "^18.0.0",
+    "react": "18.2.0",
     "react-ace": "^6.5.0",
     "react-contenteditable": "^3.2.6",
-    "react-dom": "^18.0.0",
+    "react-dom": "18.2.0",
     "react-keyboard-event-handler": "^1.4.1"
   },
   "devDependencies": {
     "@types/classnames": "^2.2.10",
     "@types/lodash": "^4.14.161",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
+    "@types/react": "18.2.67",
+    "@types/react-dom": "18.2.22",
     "@vitejs/plugin-react": "3.1.0",
     "concurrently": "^4.1.0",
     "vite": "4.1.1",
