+++ apps/evans/components/pages/explore/BookSlider.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React, { useState, useEffect } from 'react';
 import cx from 'classnames';
 import { Swipeable } from 'react-swipeable';
