+++ apps/evans/components/pages/explore/TimelineBlock.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React, { useState } from 'react';
 import cx from 'classnames';
 import NextBgImage from 'next-bg-image';
