+++ apps/evans/mdx/quakers.en.mdx
@@ -19,7 +19,7 @@ description:
   tracks={[
     {
       title: `Introduction to the Early Quakers`,
-      mp3Url: `http://flp-assets.nyc3.digitaloceanspaces.com/static/intro-to-early-quakers.en.mp3`,
+      mp3Url: `https://flp-assets.nyc3.digitaloceanspaces.com/static/intro-to-early-quakers.en.mp3`,
       duration: 1215,
     },
   ]}
