+++ apps/evans/mdx/quakers.es.mdx
@@ -22,7 +22,7 @@ description:
   tracks={[
     {
       title: `¿Quienes Eran Los Primeros Cuáqueros?`,
-      mp3Url: `http://flp-assets.nyc3.digitaloceanspaces.com/static/intro-to-early-quakers.es.mp3`,
+      mp3Url: `https://flp-assets.nyc3.digitaloceanspaces.com/static/intro-to-early-quakers.es.mp3`,
       duration: 1593,
     },
   ]}
