+++ apps/evans/pages/_app.tsx
@@ -3,10 +3,16 @@ import '@/styles/globals.css';
 import '@/styles/cover.css';
 import '@/styles/fontawesome.css';
 import Chrome from '@/components/chrome/Chrome';
+import Footer from '@/components/chrome/Footer';
 
 const App: React.FC<AppProps> = ({ Component, pageProps }) => (
   <Chrome>
-    <Component {...pageProps} />
+    <div className="pt-[70px] min-h-screen flex flex-col">
+      <main>
+        <Component {...pageProps} />
+      </main>
+      <Footer />
+    </div>
   </Chrome>
 );
