+++ apps/jones/package.json
@@ -37,9 +37,9 @@
     "lodash.get": "^4.4.2",
     "moment": "^2.24.0",
     "re-resizable": "^4.11.0",
-    "react": "^18.0.0",
+    "react": "18.2.0",
     "react-ace": "^6.3.2",
-    "react-dom": "^18.0.0",
+    "react-dom": "18.2.0",
     "react-keyboard-event-handler": "^1.4.1",
     "react-redux": "^6.0.0",
     "react-sizeme": "^2.5.2",
@@ -58,8 +58,8 @@
     "@types/lodash.defer": "4.1.7",
     "@types/lodash.get": "^4.4.6",
     "@types/node": "^11.11.3",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
+    "@types/react": "18.2.67",
+    "@types/react-dom": "18.2.22",
     "@types/react-redux": "^6.0.0",
     "@types/uuid": "^3.4.4",
     "netlify-cli": "12.14.0",
