+++ apps/native/package.json
@@ -130,7 +130,7 @@
     "@types/lodash.omit": "^4.5.6",
     "@types/lodash.throttle": "4.1.7",
     "@types/node": "18.14.2",
-    "@types/react": "^18.0.24",
+    "@types/react": "18.2.67",
     "@types/react-native-base64": "0.2.0",
     "@types/react-native-vector-icons": "6.4.13",
     "@types/react-redux": "7.1.23",
