+++ apps/poster/package.json
@@ -22,12 +22,12 @@
     "@friends-library/theme": "workspace:*",
     "@friends-library/types": "workspace:*",
     "@types/classnames": "^2.2.11",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
+    "@types/react": "18.2.67",
+    "@types/react-dom": "18.2.22",
     "@vitejs/plugin-react": "3.1.0",
     "classnames": "^2.2.6",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "tailwindcss": "3.3.1",
     "vite": "4.2.1"
   },
