+++ apps/storybook/package.json
@@ -14,15 +14,15 @@
   },
   "dependencies": {
     "@types/node": "18.15.11",
-    "@types/react": "18.0.33",
-    "@types/react-dom": "18.0.11",
+    "@types/react": "18.2.67",
+    "@types/react-dom": "18.2.22",
     "autoprefixer": "10.4.14",
     "next": "13.3.0",
     "postcss": "8.4.21",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "tailwindcss": "3.3.1",
-    "typescript": "5.0.2"
+    "typescript": "5.4.2"
   },
   "devDependencies": {
     "@friends-library/cover-component": "workspace:*",
