+++ apps/styleguide/package.json
@@ -17,16 +17,16 @@
     "@friends-library/doc-css": "workspace:*",
     "@friends-library/evaluator": "workspace:*",
     "@friends-library/types": "workspace:*",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "styled-components": "^4.0.2"
   },
   "devDependencies": {
     "@types/glob": "^7.1.3",
     "@types/lodash": "^4.14.168",
     "@types/lodash.throttle": "4.1.7",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
+    "@types/react": "18.2.67",
+    "@types/react-dom": "18.2.22",
     "@types/styled-components": "4.1.8",
     "@vitejs/plugin-react": "3.1.0",
     "chokidar": "^3.4.2",
