+++ libs-ts/doc-manifests/package.json
@@ -24,8 +24,8 @@
     "lodash.mapvalues": "^4.6.0",
     "lodash.pickby": "^4.6.0",
     "moment": "^2.29.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "roman-numerals": "^0.3.2",
     "strip-indent": "^3.0.0",
     "x-ts-utils": "^1.2.0"
@@ -33,8 +33,8 @@
   "devDependencies": {
     "@types/lodash.mapvalues": "^4.6.6",
     "@types/lodash.pickby": "^4.6.6",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
+    "@types/react": "18.2.67",
+    "@types/react-dom": "18.2.22",
     "@types/roman-numerals": "^0.3.0"
   }
 }
